+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Enable passing `string`s to `addPeerToAddressBook`.
-- Use `addPeerToAddressBook` in examples.
+- Use `addPeerToAddressBook` in examples and usage doc.
 
 ## [0.1.0] - 2021-05-12