+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.2.0] - 2021-05-14
+
 ### Added
 - `WakuRelay.getPeers` method.
 - Use `WakuRelay.getPeers` in web chat app example to disable send button.
@@ -27,5 +29,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [ReactJS Chat App example](./examples/web-chat).
 - [Typedoc Documentation](https://status-im.github.io/js-waku/docs).
 
-[Unreleased]: https://github.com/olivierlacan/keep-a-changelog/compare/v0.1.0...HEAD
-[0.1.0]: https://github.com/olivierlacan/keep-a-changelog/compare/f46ce77f57c08866873b5c80acd052e0ddba8bc9..v0.1.0
+[Unreleased]: https://github.com/status-im/js-waku/compare/v0.2.0...HEAD
+[0.2.0]: https://github.com/status-im/js-waku/compare/v0.1.0...v0.2.0
+[0.1.0]: https://github.com/status-im/js-waku/compare/f46ce77f57c08866873b5c80acd052e0ddba8bc9...v0.1.0